Version 0.9.7
Cosmetics/Features/Code:
*split ChessBoardView extending JPanel and representing a chessboard out
 of ChessGuiView as seperate class to be used in board editor
*chessGuiView no longer extends JFrame
*implemented board editor: new classes BoardEditorController and BoardEditorView
*new constructors in board, chessgame, gamecontroller to pass 
information from board editor
*deepcopy methods in Coordinate and Piece
*board now sets most fields in createPieceFields (king positions, pawn struct, piecelists) (was createPieceLists)  
*changed depth field in transTableEntry from byte to int
*removed whiteCastled and blackCastled flags in board class 

Bugs:
*fastRepetitionCheck now checks every other move (instead of every 4th)
 to see if a position is repeated
*pawnBonus and rookBonus took ownColor when checking for isolated pawn
 and open file, corrected to color of the pawn

AI:
*optimized code in pieceValue() function (all bonuses are added to value once instead of separately), removed check for rook move without castled first
*optimized bonus functions: king, bishop,
*split position matrices in respective colors to improve performance 
(only look-up necessary, no operatiosn needed to mirror them for  white)
*increased value for castling fields in king middlegame matrices
*implemented iterative deepening
*major rewrite of builtTree: from minmax to negamax algorithm (cleaner, shorter code)
*removed bestValue field (alpha saves best value), bestMove is now set 
when gameValue > alpha is checked
*removed alpha in enum evaluationflag
*if maxSearchDepth is reached, quiescenceSearch is returned immediately
 ( no storing of position in transTable)
*implemented principal variation search
*removed transposition table support from quiescence search
*removed KING_MOVE_NO_CASTLE_MALUS, CASTLINGBONUS, ROOK_MOVE_NO_CASTLE_MALUS. AI should be motivated to castle by position matrices and king safety
*increased PAWN_DOUBLE_MALUS and PAWN_ISOLATED_MALUS to 15
*AI know takes shortest way to mate, implemented constant MATEVALUE and replaced all Integer.MAX_VALUE with it,
	AI analytics now show gamevalue as "mate in x" if ai sees mate
*combined getNextMove and findNextMove
*reversed if statement of leaf and inner nodes
*implemented null move pruning
*rewrote sorting: custom sort method instead of comparator for normal sorting. Sorts by takeList(MVVLVA) + killermoves + all other moves, 
should improve performance
*got rid of depth field in chessTreeNode
*rewrote depth handling to standard in literature (depth is given as parameter when building tree)
*implemented a check for quiescence explosion into iterative deepening search

+++ src/chess/ai/TransTableEntry.java
@@ -6,7 +6,6 @@
 package chess.ai;
 
 import chess.move.Move;
-import java.util.logging.Logger;
 
 /**
  *
@@ -16,12 +15,12 @@ public class TransTableEntry {
     
     private final long zobristKey;
     private final int value;
-    private final byte depth;    
+    private final int depth;    
     private final Move bestMove;
     private final EvaluationFlag EvaluationFlag;
     private boolean oldFlag;
     
-    public TransTableEntry(long zobristKey, int value, byte depth, Move bestMove, EvaluationFlag flag) {
+    public TransTableEntry(long zobristKey, int value, int depth, Move bestMove, EvaluationFlag flag) {
         this.zobristKey = zobristKey;
         this.value = value;
         this.depth = depth;
@@ -38,7 +37,7 @@ public class TransTableEntry {
         return value;
     }
 
-    public byte getDepth() {
+    public int getDepth() {
         return depth;
     }
